+++ .php-cs-fixer.php
@@ -1,0 +1,15 @@
+<?php
+
+$finder = PhpCsFixer\Finder::create()
+    ->in(__DIR__)
+    ->exclude('vendor')
+;
+
+$config = new PhpCsFixer\Config();
+
+return $config->setRules([
+    '@PhpCsFixer' => true,
+    'php_unit_test_class_requires_covers' => false,
+])
+    ->setFinder($finder)
+;