+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 # Changelog
 All Notable changes to `oauth2-microsoft-graph` will be documented in this file
 
+## 1.0.0 - 2022-11-13
+
+### Added
+- PHP-CS-Fixer linter
+- PHPStan linter
+- Support for PHP 8.1
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Support for PHP < 8.0
+- Usage of Mockery in tests
+
+### Security
+- Nothing
+
 ## 0.1.0 - 2021-05-27
 
 ### Added